Move scalar wrapper.
