+++ cpp/tests/binaryop/integration/assert-binops.h
@@ -22,7 +22,7 @@
 
 #include <tests/binaryop/util/operation.h>
 #include <tests/utilities/legacy/column_wrapper.cuh>
-#include <tests/utilities/scalar_wrapper.cuh>
+#include <tests/utilities/legacy/scalar_wrapper.cuh>
 #include <gtest/gtest.h>
 
 namespace cudf {
