+++ cpp/tests/binaryop/unit/binop-verify-input-test.cpp
@@ -18,7 +18,7 @@
  */
 
 #include <tests/utilities/legacy/column_wrapper.cuh>
-#include <tests/utilities/scalar_wrapper.cuh>
+#include <tests/utilities/legacy/scalar_wrapper.cuh>
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
 #include <cudf/binaryop.hpp>
 #include <gtest/gtest.h>
