+++ cpp/tests/copying/copy_range_tests.cu
@@ -17,7 +17,7 @@
 #include <cudf/copying.hpp>
 #include <tests/utilities/legacy/column_wrapper.cuh>
 #include <tests/utilities/legacy/column_wrapper_factory.hpp>
-#include <tests/utilities/scalar_wrapper.cuh>
+#include <tests/utilities/legacy/scalar_wrapper.cuh>
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
 #include <tests/utilities/nvcategory_utils.cuh>
