+++ cpp/tests/filling/filling_tests.cu
@@ -16,7 +16,7 @@
 
 #include <cudf/filling.hpp>
 #include <tests/utilities/legacy/column_wrapper.cuh>
-#include <tests/utilities/scalar_wrapper.cuh>
+#include <tests/utilities/legacy/scalar_wrapper.cuh>
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
 
 template <typename T>
