+++ cpp/tests/filling/repeating_tests.cu
@@ -17,7 +17,7 @@
 #include <cudf/filling.hpp>
 #include <tests/utilities/legacy/column_wrapper.cuh>
 #include <tests/utilities/legacy/column_wrapper_factory.hpp>
-#include <tests/utilities/scalar_wrapper.cuh>
+#include <tests/utilities/legacy/scalar_wrapper.cuh>
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
 
 #include <numeric>
