+++ cpp/tests/iterator/iterator_test.cu
@@ -28,7 +28,7 @@
 
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
 #include <tests/utilities/legacy/column_wrapper.cuh>
-#include <tests/utilities/scalar_wrapper.cuh>
+#include <tests/utilities/legacy/scalar_wrapper.cuh>
 #include <utilities/device_operators.cuh>
 
 #include <thrust/transform.h>
