+++ cpp/tests/quantiles/quantiles_test.cu
@@ -19,7 +19,7 @@
 
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
 #include <tests/utilities/legacy/column_wrapper.cuh>
-#include <tests/utilities/scalar_wrapper.cuh>
+#include <tests/utilities/legacy/scalar_wrapper.cuh>
 
 #include <utilities/cudf_utils.h>
 #include <utilities/error_utils.hpp>
