+++ cpp/tests/reductions/reduction_tests.cu
@@ -32,7 +32,7 @@
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
 #include <tests/utilities/legacy/cudf_test_utils.cuh>
 #include <tests/utilities/legacy/column_wrapper.cuh>
-#include <tests/utilities/scalar_wrapper.cuh>
+#include <tests/utilities/legacy/scalar_wrapper.cuh>
 
 template <typename T>
 std::vector<T> convert_values(std::vector<int> const & int_values)
