+++ cpp/tests/replace/replace-nulls_tests.cu
@@ -20,7 +20,7 @@
 #include <utilities/error_utils.hpp>
 
 #include <tests/utilities/legacy/column_wrapper.cuh>
-#include <tests/utilities/scalar_wrapper.cuh>
+#include <tests/utilities/legacy/scalar_wrapper.cuh>
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
 #include <tests/utilities/legacy/cudf_test_utils.cuh>
 #include <cudf/cudf.h>
