+++ cpp/tests/utilities/legacy/scalar_wrapper.cuh
@@ -18,7 +18,7 @@
 #define SCALAR_WRAPPER_H
 
 #include <cudf/cudf.h>
-#include "legacy/cudf_test_utils.cuh"
+#include <tests/utilities/legacy/cudf_test_utils.cuh>
 #include <rmm/rmm.h>
 #include <utilities/bit_util.cuh>
 #include <cudf/utilities/legacy/type_dispatcher.hpp>
